fix: replace unsafe numeric casts with checked conversions
Fixes #679, fixes #689.

- weather.rs: last_check_hour field/accessor changed from Option<u32> to
  Option<i64>; tick() now computes current_hour via now.timestamp()/3600
  without the as-u32 cast. For game dates in 1820 the timestamp is a large
  negative i64 — the previous as u32 silently wrapped, producing a wrong
  hour value and breaking the once-per-hour dedup guard.
- debug_snapshot.rs: WeatherEngineSnapshot.last_check_hour updated to match.
- pipeline.rs: three usize→u32 casts guarded with u32::try_from().expect().
- connections.rs: debug-log best_dist cast changed from as u32 to as u64
  (distances are always positive; u64 avoids truncation on large values).

+++ crates/parish-world/src/weather.rs
@@ -96,7 +96,7 @@ pub struct WeatherEngine {
     /// Minimum duration in game-hours before a transition is allowed.
     min_duration_hours: f64,
     /// Game-hour of the last transition check (to avoid multiple checks per hour).
-    last_check_hour: Option<u32>,
+    last_check_hour: Option<i64>,
     /// Ring buffer of recent weather transitions: (timestamp, new_weather).
     history: VecDeque<(DateTime<Utc>, Weather)>,
 }
@@ -142,7 +142,7 @@ impl WeatherEngine {
     }
 
     /// Returns the game-hour of the last transition check, or `None` if unchecked.
-    pub fn last_check_hour(&self) -> Option<u32> {
+    pub fn last_check_hour(&self) -> Option<i64> {
         self.last_check_hour
     }
 
@@ -160,7 +160,7 @@ impl WeatherEngine {
         season: Season,
         rng: &mut impl Rng,
     ) -> Option<Weather> {
-        let current_hour = now.timestamp() as u32 / 3600;
+        let current_hour = now.timestamp() / 3600;
 
         // Only check once per game-hour
         if self.last_check_hour == Some(current_hour) {
